Drop the combined grep package.
